requirements.txt
numpy==1.14.5
Pillow==5.2.0
scikit-learn==0.19.1
scipy==1.1.0
sklearn==0.0
